Added some factory classes for spawning forms.
Moved the Board watching code into it's own service so it's not bound to a single form.

BoardView can now be launched independently of other views.

Created a web cam selector dialog.

Kernel is not set up in program.cs

State.cs - had to add some fixes to prevent crashing